[Refactor] 스킬 네트워크 엔트리 타입 파일 분리
