+++ RAW_unity/Assets/_RAW/Scripts/Network/Skills/DefaultSkillLoadoutEntry.cs
@@ -1,0 +1,15 @@
+using System;
+using UnityEngine;
+
+namespace RAW.Network
+{
+	[Serializable]
+	public class DefaultSkillLoadoutEntry
+	{
+		[SerializeField] private KeyMapping slot;
+		[SerializeField] private SkillSpec skill;
+
+		public KeyMapping Slot => slot;
+		public SkillSpec Skill => skill;
+	}
+}
