+++ RAW_unity/Assets/_RAW/Scripts/Network/Skills/DefaultSkillLoadoutEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e18e1340a9147432088e3cd7460c66c1
