+++ RAW_unity/Assets/_RAW/Scripts/Network/Skills/NetworkSkillCooldownEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 141a834d15da34bc3a143a9f20a740b1
