+++ RAW_unity/Assets/_RAW/Scripts/Network/Skills/NetworkSkillLoadoutEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 33e55469c784f41e39b9b2098b420e3a
